initial commit
the gen.tmpl file is a template for a perl script that gets generated
by the shell script make_comic_html.sh - it depends on the external
utilities unrar, unzip and a working bash and perl interpreter.  The
jpg files are generic logos for navigation - anything can be used in
their place.